Amélioration design site, Gestion urls medias (Dropbox, Zip ou tar.gz...)

+++ escapad_formulaire/views.py
@@ -39,6 +39,7 @@ import urllib2
 #RANDOM
 import string
 import random
+import re
 
 #################################
 #                               #
@@ -67,6 +68,7 @@ def form_upload(request):
 
         modulesData=[]
         mediasData=[]
+        mediasType=[]
         nbModule=request.POST.get("nb_module")
 
         #Go through each modules to get the md and media data
@@ -79,7 +81,14 @@ def form_upload(request):
             modulesData.append(moduleData)
             mediasData.append(mediaData)
 
-        zip=cn.generateArchive(modulesData,mediasData,homeData,titleData,logoData,repoDir,outDir,baseUrl)
+            # Specify if the media is empty or not (tar.gz or empty)
+            #TODO No zip supported for now (Improve?)
+            if mediaData:
+                mediasType.append('application/octet-stream')
+            else:
+                mediasType.append('None')
+
+        zip=cn.generateArchive(modulesData,mediasData,mediasType,homeData,titleData,logoData,repoDir,outDir,baseUrl)
 
         sauvegarde = True
 
@@ -150,6 +159,7 @@ def form_upload_eth(request):
 
         modulesData=[]
         mediasData=[]
+        mediasType=[]
         nbModule=request.POST.get("nb_module")
 
         #Go through each module to get the md and media data
@@ -162,12 +172,18 @@ def form_upload_eth(request):
             moduleData= StringIO.StringIO(response.read())
 
             nomMedia="media_"+str(i)
-            mediaData=formMod.cleaned_data[nomMedia]
+            mediaData=request.POST.get(nomMedia)
 
             modulesData.append(moduleData)
             mediasData.append(mediaData)
 
-        zip=cn.generateArchive(modulesData,mediasData,homeData,titleData,logoData,repoDir,outDir,baseUrl)
+            #Specify if the media is empty or not (tar.gz or empty)
+            if mediaData:
+                mediasType.append('application/octet-stream')
+            else:
+                mediasType.append('None')
+
+        zip=cn.generateArchive(modulesData,mediasData,mediasType,homeData,titleData,logoData,repoDir,outDir,baseUrl)
 
         sauvegarde = True
 
@@ -238,7 +254,10 @@ def id_generator(size=6, chars=string.ascii_lowercase + string.digits):
 # View resuming the user's courses
 # The user can either access an existing course, or create a new one
 def mes_cours(request):
-    print request
+
+    if not request.user.is_authenticated:
+        return redirect(connexion)
+
     myUser=request.user
     profil= Profil.objects.get(user=myUser)
     form = CreateNew(request.POST or None)
@@ -246,7 +265,7 @@ def mes_cours(request):
     if form.is_valid() :
         url_home='home-'+id_generator()
         id_cours=id_generator()
-        cours=Cours(id_cours=id_cours, nom_cours=form.cleaned_data['nom'], nb_module=1, url_home=url_home)
+        cours=Cours(id_cours=id_cours, nom_cours=form.cleaned_data['nom'], nb_module=0, url_home=url_home)
         cours.save()
         profil.cours.add(cours)
         #url_home='http://193.51.236.202:9001/p/'+url_home
@@ -256,20 +275,33 @@ def mes_cours(request):
         'form' : form
     })
 
-# View
+# View showing the information of a course
 def cours(request, id_cours):
 
-    cours= Cours.objects.get(id_cours=id_cours)
+    cours = Cours.objects.get(id_cours=id_cours)
     form = CreateNew(request.POST or None)
-    if form.is_valid() :
-        url='module-'+id_generator()
-        module=Module(nom_module=form.cleaned_data['nom'], url=url, cours=cours)
-        module.save()
-
-
     form2 = UploadFormEth(request.POST or None, request.FILES or None)
-    if form2.is_valid() :
+    form3 = SearchUser(request.POST or None)
 
+    # Adding a module to the course
+    if form.is_valid() :
+        url = 'module-'+id_generator()
+        module = Module(nom_module=form.cleaned_data['nom'], url=url, cours=cours)
+        module.save()
+        cours.nb_module=cours.nb_module+1
+        cours.save()
+
+
+    userFound = False
+    # Adding a user to the course
+    if form3.is_valid() :
+        user = User.objects.get(username = form3.cleaned_data['user'])
+        profil = user.profil
+        profil.cours.add(cours)
+        userFound = True
+
+    # Generating the course
+    if form2.is_valid() :
         repoDir=settings.BASE_DIR
     	outDir=settings.BASE_DIR
     	baseUrl=settings.BASE_DIR
@@ -277,7 +309,6 @@ def cours(request, id_cours):
         titleData=form2.cleaned_data["nom_cours"]
         logoData=form2.cleaned_data["logo"]
 
-        #print cours.url_home
         url_home='http://193.51.236.202:9001/p/'+cours.url_home+'/export/txt'
 
 
@@ -289,22 +320,32 @@ def cours(request, id_cours):
 
         modulesData=[]
         mediasData=[]
+        mediasType=[]
 
+        #for each modules from the course
         for module in cours.module_set.all():
+            # get the pad content
             url_module= 'http://193.51.236.202:9001/p/'+module.url+'/export/txt'
             response = urllib2.urlopen(url_module)
             moduleData = StringIO.StringIO(response.read())
-            #print moduleData.read()
             modulesData.append(moduleData)
-            mediasData.append('')
+
+            # open the dropbox link to get the archive
+            url_media=module.url_media
+            response = urllib2.urlopen(url_media)
+
+            # get the media archive type of content (we try to get either application/zip or application/octet-stream)
+            http_message=response.info()
+            full=http_message.type
+            mediasType.append(full)
+
+            # read the response and add it to the media datas
+            mediaData = StringIO.StringIO(response.read())
+            mediasData.append(mediaData)
 
 
+        zip=cn.generateArchive(modulesData,mediasData,mediasType,homeData,titleData,logoData,repoDir,outDir,baseUrl)
 
-        #homeData=""
-
-        zip=cn.generateArchive(modulesData,mediasData,homeData,titleData,logoData,repoDir,outDir,baseUrl)
-
-        sauvegarde = True
 
         response= HttpResponse(zip)
         response['Content-Type'] = 'application/octet-stream'
@@ -313,40 +354,88 @@ def cours(request, id_cours):
 
 
     cours=Cours.objects.get(id_cours=id_cours)
-    print cours.nom_cours
+
     return render(request, 'escapad_formulaire/cours.html', {
         'cours' : cours,
         'form' : form,
-        'form2' : form2
+        'form2' : form2,
+        'form3' : form3,
+        'userFound' : userFound
+
     })
 
 # View form for creating the home file
 def cours_edition(request, id_cours ,url):
 
+    if not request.user.is_authenticated:
+        return redirect(connexion)
 
-    #url_home='home-'+id_generator()
+    form_media = MediaForm(request.POST or None)
+    cours = Cours.objects.get(id_cours=id_cours)
     full_url='http://193.51.236.202:9001/p/'+url
 
+    #if we changed the media url
+    if form_media.is_valid() :
+        res_url=form_media.cleaned_data['url_media']
+        #if the url if from dropbox, we change dl=0 to dl=1, in order to create a direct download link
+        # (not provided by dropbox directly), I guess it's better to do it rather than asking the user.
+        if re.match(r"^.*dropbox.*dl=",res_url):
+            res_url=re.sub(r"^(?P<debut>.*)dl=[0-9](?P<fin>.*)$", r"\g<debut>dl=1\g<fin>", res_url)
 
+        #If we modify the home page
+        if re.match(r"^home",url):
+            cours.url_media=res_url
+            cours.save()
+        #If we modify a module page
+        elif re.match(r"^module",url):
+            module=Module.objects.get(url=url)
+            module.url_media=res_url
+            module.save()
 
+    url_media=''
+    #get the url_media to print on the view
+    #home page
+    if re.match(r"^home",url):
+        url_media= cours.url_media
+    #module page
+    elif re.match(r"^module",url):
+        module=Module.objects.get(url=url)
+        url_media= module.url_media
 
     return render(request, 'escapad_formulaire/form_edition.html', {
         'id_cours': id_cours,
         'url': url,
         'full_url': full_url,
-        'sauvegarde': False
+        'url_media': url_media,
+        'form_media': form_media,
     })
 
-
-
+# Delete a module from a course
+# Irreversible task
 def delete_module(request, id_cours, url):
+    if not request.user.is_authenticated:
+        return redirect(connexion)
+
     module= Module.objects.get(url=url)
     module.delete()
+    c= Cours.objects.get(id_cours=id_cours)
+    c.nb_module=c.nb_module-1
+    c.save()
     return redirect(cours,id_cours=id_cours)
 
+# Delete a course, then redirect the user to its course list page
+# Note: If a course is shared by several user, the course will not be deleted,
+# we will just remove the ManyToManyField relation between the user and the course
 def delete_course(request, id_cours):
+    if not request.user.is_authenticated:
+        return redirect(connexion)
+
     course= Cours.objects.get(id_cours=id_cours)
-    course.delete()
+    if course.profil_set.all().count == 1:
+        course.delete()
+    else:
+        profil=request.user.profil
+        course.profil_set.remove(profil)
     return redirect(mes_cours)
 
 
@@ -380,7 +469,6 @@ def connexion(request):
 # The form create a user (Django native class), and associate him to a profile, which will contain the user's projects
 def inscription(request):
     error = False
-    print "oooooooooooooooo"
     if request.method == "POST":
         form = CreateUserForm(request.POST)
         print form.is_valid()
